diffraction: extract calibration out of BraggPeaksPolymer
None of this was polymer-specific. Measured inside exact AST method boundaries,
the calibration cluster referenced `self` only to reach a Dataset4dstem, a peak
Vector, a device string, and a slot to deposit diagnostics into -- all general
core types. The two ellipse fitters take a plain 2D array and a centre, so they
apply to any diffuse or crystalline ring from any instrument; `_fit_origin_roi`
did not reference `self` at all.

New `ellipse_fitting.py` holds `fit_ellipse_from_ring` (angular variance) and
`fit_ellipse_from_ridge` (joint centre + ellipse), returning an `EllipseFit`
dataclass. Returning the fit instead of assigning `self.ellipse_fit_diagnostics`
was what actually completed the extraction: that assignment was the only thing
tying 837 lines of arithmetic to the class. `EllipseFit.diagnostics` rebuilds
the flat mapping the class has always published, so `bp.ellipse_fit_diagnostics`
keeps its exact key set -- pinned by new tests for both methods.

Origin/descan work moves to `polar_transform.py`, which already owned three
origin finders, rather than to a third location: `find_central_beams_4d`,
`find_central_beams_from_peaks_4d`, `fit_origin_roi`, `centered_dp_mean`. The
dataset, peaks and device are now parameters. `find_central_beams_4d` stays on
the class as a forwarding wrapper since it is public API; every internal caller
already used keyword arguments.

BraggPeaksPolymer: 79 -> 74 methods, 6798 -> 5691 lines. bragg_peaks.py drops
below 8k. test_ellipse_ring_fit.py no longer constructs the class via
object.__new__ to reach the fitters; the one remaining use there legitimately
exercises preprocess argument validation.

Tests: 488 passed, 10 skipped across the full suite (102 in tests/diffraction).
Verified separately that both fitters still recover a synthetic ellipse
(true ratio 1.111 -> 1.105 ring / 1.117 ridge; true theta 35 deg -> 36.0 / 35.3).
