+++ src/quantem/diffraction/bragg_peaks.py
@@ -27,10 +27,14 @@ from quantem.diffraction.polymer_normalization import (
 from quantem.core.datastructures import Vector
 from quantem.core.visualization import show_2d
 from quantem.diffraction.polar_transform import (
+    centered_dp_mean,
+    find_central_beams_4d,
+    fit_origin_roi,
     find_origin as find_origin_angular_uniformity,
     polar_transform as karen_polar_transform,
     polar_transform_peaks as karen_polar_transform_peaks,
 )
+from quantem.diffraction.ellipse_fitting import fit_ellipse_from_ridge, fit_ellipse_from_ring
 from quantem.diffraction.peak_detection import detect_blobs, find_central_beam_from_peaks
 from quantem.diffraction.orientation_correlation import (
     calculate_orientation_correlation as _calculate_orientation_correlation,
@@ -1877,7 +1881,7 @@ class BraggPeaksPolymer(AutoSerialize):
             if estimate_descan or estimate_detector_rotation or center_source == "descan":
                 import torch
 
-                fitted = self._fit_origin_roi(measured, roi, fit_method=descan_fit_method)
+                fitted = fit_origin_roi(measured, roi, fit_method=descan_fit_method)
                 self.descan_origin = np.moveaxis(fitted, -1, 0)  # (2, Ry, Rx)
                 results["descan_origin"] = self.descan_origin
 
@@ -1927,17 +1931,17 @@ class BraggPeaksPolymer(AutoSerialize):
         self.dp_mean_centered = None
         self.ellipse_fit_diagnostics = None
         if fit_ellipse:
-            self.dp_mean_centered = self._centered_dp_mean(
-                self.image_centers, com_model=com_model
+            self.dp_mean_centered = centered_dp_mean(
+                self._dataset_cartesian, self.image_centers, com_model=com_model
             )
             Qy, Qx = self._dataset_cartesian.shape[-2:]
             center = ((Qy - 1) / 2.0, (Qx - 1) / 2.0)  # _centered_dp_mean puts the beam here
             fit_function = (
-                self._fit_ellipse_from_ridge
+                fit_ellipse_from_ridge
                 if ellipse_fit_method == "ridge"
-                else self._fit_ellipse_from_ring
+                else fit_ellipse_from_ring
             )
-            a_axis, b_axis, theta_deg, ring_band = fit_function(
+            ellipse_fit = fit_function(
                 self.dp_mean_centered,
                 center,
                 radial_min=ellipse_radial_min,
@@ -1946,15 +1950,11 @@ class BraggPeaksPolymer(AutoSerialize):
                 show=show,
                 verbose=verbose,
             )
-            self.ellipse_params = (float(a_axis), float(b_axis), float(theta_deg))
-            refined_center = self.ellipse_fit_diagnostics.get(
-                "center_refined", center
-            )
-            self.ellipse_center = tuple(float(value) for value in refined_center)
-            if (
-                ellipse_fit_method == "ridge"
-                and self.ellipse_fit_diagnostics["accepted"]
-            ):
+            ring_band = ellipse_fit.ring_band
+            self.ellipse_params = ellipse_fit.params
+            self.ellipse_fit_diagnostics = ellipse_fit.diagnostics
+            self.ellipse_center = ellipse_fit.center_refined
+            if ellipse_fit_method == "ridge" and ellipse_fit.accepted:
                 center_delta = np.asarray(self.ellipse_center) - np.asarray(center)
                 self.image_centers = np.asarray(
                     self.image_centers, dtype=float
@@ -2001,982 +2001,6 @@ class BraggPeaksPolymer(AutoSerialize):
                 print(f"  sampling       {self.sampling_inv_A:.5g} 1/A per pixel")
 
         return results
-
-    def _fit_origin_roi(self, measured, mask, fit_method="plane", clip_sigma=5.0, n_iter=2):
-        """Fit the smooth CoM-origin background over the ROI (scan mask) only.
-
-        ``measured`` is the per-pattern CoM origin ``(Ry, Rx, 2)``; ``mask`` is the
-        ``(Ry, Rx)`` scan ROI. Fitting over the whole scan lets out-of-ROI patterns
-        (vacuum/substrate, whose CoM is meaningless) drag the plane, leaving a uniform
-        residual inside the ROI. Each component is fit independently: ``"plane"`` does an
-        ordinary least-squares ``z = a + b*row + c*col`` with ``n_iter`` sigma-clip passes
-        to reject hot/dead-pixel CoM outliers; ``"constant"`` uses the ROI mean. Returns
-        the fitted field ``(Ry, Rx, 2)`` evaluated at every scan position.
-        """
-        measured = np.asarray(measured, dtype=float)
-        Ry, Rx, ncomp = measured.shape
-        m0 = np.asarray(mask, dtype=bool)
-        yy, xx = np.mgrid[0:Ry, 0:Rx].astype(float)
-        fitted = np.empty_like(measured)
-        for c in range(ncomp):
-            z = measured[..., c]
-            if fit_method == "constant":
-                ref = z[m0] if m0.any() else z
-                fitted[..., c] = float(np.nanmean(ref))
-                continue
-            use = m0 & np.isfinite(z)
-            plane = np.full((Ry, Rx), float(np.nanmean(z[use])) if use.any() else 0.0)
-            for _ in range(max(1, n_iter)):
-                if int(use.sum()) < 3:
-                    break
-                A = np.stack([np.ones(int(use.sum())), xx[use], yy[use]], axis=1)
-                coef, *_ = np.linalg.lstsq(A, z[use], rcond=None)
-                plane = coef[0] + coef[1] * xx + coef[2] * yy
-                resid = z - plane
-                s = float(np.std(resid[use]))
-                if s == 0:
-                    break
-                new_use = m0 & np.isfinite(z) & (np.abs(resid) < clip_sigma * s)
-                if int(new_use.sum()) == int(use.sum()) or int(new_use.sum()) < 8:
-                    break
-                use = new_use
-            fitted[..., c] = plane
-        return fitted
-
-    def _centered_dp_mean(self, image_centers, com_model=None):
-        """Mean diffraction pattern with every pattern shifted so its central beam
-        lands at the detector center.
-
-        Averaging the raw patterns smears the diffraction ring by the descan drift and
-        leaves the central beam off-center, which biases an ellipse fit toward the
-        central beam. Aligning each pattern first yields a sharp, concentric ring.
-
-        When a ``CenterOfMassOriginModel`` is available (it holds the 4D tensor on-device),
-        each ROI pattern is SUB-PIXEL shifted (bilinear) by ``image_centers`` -- the
-        angular-uniformity BEAM center -- so the beam lands exactly on the detector center,
-        and the mean is accumulated in batches. We center by ``image_centers`` rather than
-        the CoM/descan origin because the CoM is the centroid of the whole pattern: any
-        ring or background asymmetry pulls it a few px off the actual beam, which would
-        leave the beam off-center in the mean. Otherwise (no CoM model) we fall back to
-        translating the plain mean DP by the average center offset.
-
-        Sub-pixel matters: plain integer rolls leave a per-pattern residual of up to 0.5 px
-        that does NOT average out when the center spread is narrow (all patterns round to
-        the same integer), so the mean beam ends up biased off-center by a fraction of a
-        pixel. Bilinear splatting removes that bias.
-
-        We deliberately avoid ``CenterOfMassOriginModel.shift_origin_to`` here: it
-        materialises a full second copy of the 4D stack *plus* a per-pattern sampling
-        grid, which OOMs on large scans. The bilinear accumulator needs only a
-        ``(Qy, Qx)`` buffer plus one batch of patterns at a time.
-        """
-        Qy, Qx = self._dataset_cartesian.shape[-2:]
-        center = ((Qy - 1) / 2.0, (Qx - 1) / 2.0)
-        ic = np.asarray(image_centers, dtype=float)  # (2, Ry, Rx); 0 outside the scan mask
-        valid = (ic[0] != 0) | (ic[1] != 0)
-        if com_model is not None and valid.any():
-            import torch
-
-            with torch.no_grad():
-                flat = com_model.tensor.reshape(-1, Qy, Qx)
-                dev = flat.device
-                # Center by the authoritative per-pattern beam centers (image_centers, from
-                # the angular-uniformity finder), NOT the CoM plane: the CoM is the centroid
-                # of the whole pattern, so ring/background asymmetry pulls it a few px off
-                # the beam, and centering by it would leave the beam off-center. Only ROI
-                # (in-mask, non-zero) patterns are averaged.
-                oy = torch.as_tensor(ic[0].ravel(), dtype=torch.float, device=dev)
-                ox = torch.as_tensor(ic[1].ravel(), dtype=torch.float, device=dev)
-                keep = (
-                    torch.as_tensor(valid.ravel(), device=dev)
-                    .nonzero(as_tuple=False)
-                    .squeeze(1)
-                )
-                sy = center[0] - oy  # continuous shift -> detector center, (y, x)
-                sx = center[1] - ox
-                acc = torch.zeros((Qy, Qx), dtype=torch.float32, device=dev)
-                batch = 256
-                for bstart in range(0, int(keep.numel()), batch):
-                    bidx = keep[bstart:bstart + batch]
-                    chunk = flat[bidx].float()                 # (b, Qy, Qx), ROI patterns only
-                    fyb = torch.floor(sy[bidx])
-                    fxb = torch.floor(sx[bidx])
-                    gy = sy[bidx] - fyb
-                    gx = sx[bidx] - fxb
-                    floor_pairs = torch.stack([fyb.long(), fxb.long()], dim=1)  # (b, 2)
-                    # bilinear weights for the 4 integer-shift corners around the fraction
-                    corner_w = {
-                        (0, 0): (1 - gy) * (1 - gx),
-                        (0, 1): (1 - gy) * gx,
-                        (1, 0): gy * (1 - gx),
-                        (1, 1): gy * gx,
-                    }
-                    # descan drift is smooth -> few distinct floor shifts per batch:
-                    # weight-sum each group, then roll each of the 4 corners.
-                    uniq, inv = torch.unique(floor_pairs, dim=0, return_inverse=True)
-                    for k in range(int(uniq.shape[0])):
-                        m = inv == k
-                        cg = chunk[m]                          # (g, Qy, Qx)
-                        fy = int(uniq[k, 0])
-                        fx = int(uniq[k, 1])
-                        for (dy, dx), wt in corner_w.items():
-                            s = (cg * wt[m][:, None, None]).sum(0)
-                            acc += torch.roll(s, shifts=(fy + dy, fx + dx), dims=(0, 1))
-                dp = (acc / max(int(keep.numel()), 1)).detach().cpu().numpy()
-            return np.asarray(dp, dtype=float)
-        # Fallback: translate the raw mean DP so the average beam center is centered.
-        # image_centers is 0 outside the scan mask (find_central_beams_4d), so average
-        # over valid (non-zero) positions only to avoid a bias toward the origin.
-        from scipy.ndimage import shift as ndi_shift
-
-        dp = np.asarray(self._dataset_cartesian.get_dp_mean().array, dtype=float)
-        valid = (image_centers[0] != 0) | (image_centers[1] != 0)
-        if not valid.any():
-            valid = np.ones_like(image_centers[0], dtype=bool)
-        dy = center[0] - float(image_centers[0][valid].mean())
-        dx = center[1] - float(image_centers[1][valid].mean())
-        return ndi_shift(dp, (dy, dx), order=1, mode="constant", cval=0.0)
-
-    def _fit_ellipse_from_ring(
-        self,
-        dp,
-        center,
-        *,
-        radial_min=None,
-        radial_max=None,
-        radial_step=1.0,
-        num_annular_bins=180,
-        ratio_range=(0.85, 1.18),
-        n_ratio=12,
-        n_theta=24,
-        refine=True,
-        max_ring_candidates=3,
-        min_fit_improvement=0.005,
-        max_fit_score=0.25,
-        min_angular_coverage=0.55,
-        device="cpu",
-        show=False,
-        verbose=False,
-    ):
-        """Fit ring ellipticity ``(a, b, theta_deg)`` by minimising the azimuthal variance
-        of a diffraction-ring annulus at a FIXED center -- Karen Ehrhardt's angular-
-        uniformity criterion (see ``quantem.diffraction.polar_transform``).
-
-        Unlike a probe-blob fit (``fit_probe_ellipse``) this samples an annulus out at the
-        ring radius and never touches the central beam, so a smeared / off-center / doubled
-        central beam does not bias the result. Only the axis ratio ``b/a`` and orientation
-        ``theta`` are identifiable from a single ring, so the returned ``(a, b)`` are
-        normalised to the ring radius (``a ~ R0``); downstream consumers (``polar_transform``
-        / ``find_central_beams_4d``) use only ``b/a`` and ``theta``.
-
-        Parameters
-        ----------
-        dp : ndarray
-            Centered mean diffraction pattern (beam at ``center``).
-        center : (float, float)
-            Fixed origin ``(y, x)`` in detector pixels.
-        radial_min, radial_max : float, optional
-            Ring band in pixels. If either is None the band is auto-detected from the
-            circular median radial profile and several prominent candidates are fitted
-            and quality-ranked.
-        ratio_range, n_ratio, n_theta, refine :
-            Coarse grid over ``b/a`` and ``theta`` (degrees), then a clipped local
-            refinement pass.
-        max_ring_candidates : int
-            Maximum prominent radial-profile peaks evaluated when the ring band is
-            selected automatically.
-        min_fit_improvement, max_fit_score, min_angular_coverage : float
-            Quality gates. Fits that do not improve held-out angular alignment, retain
-            excessive raw angular variance, lack ring coverage, or hit a ratio boundary
-            fall back to a circular correction with a warning.
-
-        Returns
-        -------
-        (a, b, theta_deg, (radial_min, radial_max))
-        """
-        from quantem.diffraction.polar_transform import polar_transform
-
-        dp = np.asarray(dp, dtype=float)
-        Qy, Qx = dp.shape
-        origin = np.asarray(center, dtype=float)
-
-        def _polar(image, ellipse_params, rmin, rmax):
-            # polar_transform returns (n_phi, n_r) when scan_pos is given.
-            return np.asarray(
-                polar_transform(
-                    image,
-                    origin_array=origin,
-                    ellipse_params=ellipse_params,
-                    num_annular_bins=num_annular_bins,
-                    radial_min=float(rmin),
-                    radial_max=float(rmax),
-                    radial_step=radial_step,
-                    scan_pos=(0, 0),
-                    device=device,
-                    show_progress=False,
-                ),
-                dtype=float,
-            )
-
-        # Log compression plus global winsorisation strongly reduces the leverage of
-        # isolated Bragg spots without erasing the broad diffuse calibration ring.
-        fit_dp = np.log1p(np.clip(dp, 0.0, None))
-        finite_fit = fit_dp[np.isfinite(fit_dp)]
-        if finite_fit.size:
-            fit_dp = np.minimum(fit_dp, np.percentile(finite_fit, 99.5))
-
-        # 1. Find several plausible diffuse-ring bands. A median angular profile is much
-        #    less likely than a mean profile to select a sparse constellation of Bragg
-        #    spots. Candidate fits are quality-ranked below rather than trusting the
-        #    single strongest radial feature.
-        r_hi = float(min(Qy, Qx) / 2.0 - 1.0)
-        explicit_band = radial_min is not None and radial_max is not None
-        candidate_bands = []
-        if radial_min is None or radial_max is None:
-            from scipy.ndimage import uniform_filter1d
-
-            prof = _polar(fit_dp, (1.0, 1.0, 0.0), 0.0, r_hi)
-            radial_profile = uniform_filter1d(np.median(prof, axis=0), size=5)
-            r_axis = np.arange(radial_profile.size) * radial_step
-            r_exclude = max(6.0, 0.06 * r_hi)
-            i0 = int(r_exclude / radial_step)
-            search_profile = radial_profile.copy()
-            search_profile[:i0] = np.min(search_profile)
-            prominence_floor = max(
-                1e-9, 0.03 * float(np.ptp(search_profile[i0:]))
-            )
-            peak_indices, properties = find_peaks(
-                search_profile,
-                prominence=prominence_floor,
-                distance=max(3, int(round(6.0 / radial_step))),
-            )
-            valid = (
-                (peak_indices >= i0)
-                & (r_axis[peak_indices] <= 0.92 * r_hi)
-            )
-            peak_indices = peak_indices[valid]
-            prominences = properties["prominences"][valid]
-            if not peak_indices.size:
-                peak_indices = np.asarray(
-                    [i0 + int(np.argmax(search_profile[i0:]))]
-                )
-                prominences = np.asarray([1.0])
-            order = np.argsort(prominences)[::-1][:max_ring_candidates]
-            for index in peak_indices[order]:
-                r0 = float(r_axis[index])
-                half = max(6.0, 0.20 * r0)
-                band_min = (
-                    max(r_exclude, r0 - half)
-                    if radial_min is None
-                    else float(radial_min)
-                )
-                band_max = (
-                    min(r_hi, r0 + half)
-                    if radial_max is None
-                    else float(radial_max)
-                )
-                if band_max > band_min:
-                    candidate_bands.append((band_min, band_max, r0))
-        else:
-            candidate_bands.append(
-                (float(radial_min), float(radial_max),
-                 0.5 * (float(radial_min) + float(radial_max)))
-            )
-
-        # Deduplicate overlapping candidates created by broad/shouldered peaks.
-        unique_bands = []
-        for band in candidate_bands:
-            if not any(abs(band[2] - other[2]) < 3.0 for other in unique_bands):
-                unique_bands.append(band)
-        candidate_bands = unique_bands
-
-        fit_angles = (np.arange(num_annular_bins) // 6) % 2 == 0
-        validation_angles = ~fit_angles
-
-        def _robust_score(ellipse_params, band, angle_mask):
-            polar = _polar(
-                fit_dp, ellipse_params, band[0], band[1]
-            )[angle_mask]
-            if not polar.size:
-                return np.inf
-            # Per-radius clipping removes angularly isolated hot pixels. Per-angle
-            # normalisation then scores radial alignment rather than polymer texture.
-            upper = np.percentile(polar, 90.0, axis=0, keepdims=True)
-            polar = np.minimum(polar, upper)
-            polar = polar - np.percentile(
-                polar, 10.0, axis=1, keepdims=True
-            )
-            polar = np.clip(polar, 0.0, None)
-            scale = np.percentile(polar, 90.0, axis=1, keepdims=True)
-            valid_scale = scale[:, 0] > 1e-9
-            if np.count_nonzero(valid_scale) < 4:
-                return np.inf
-            polar = polar[valid_scale] / (scale[valid_scale] + 1e-9)
-            reference = np.median(polar, axis=0)
-            return float(
-                np.median(np.abs(polar - reference), axis=0).sum()
-                / (np.abs(reference).sum() + 1e-9)
-            )
-
-        def _raw_score(ellipse_params, band):
-            polar = _polar(dp, ellipse_params, band[0], band[1])
-            return float(
-                polar.std(axis=0).sum()
-                / (np.abs(polar.mean(axis=0)).sum() + 1e-6)
-            )
-
-        def _angular_coverage(ellipse_params, band):
-            polar = _polar(fit_dp, ellipse_params, band[0], band[1])
-            contrast = np.percentile(polar, 95.0, axis=1) - np.percentile(
-                polar, 20.0, axis=1
-            )
-            reference = np.percentile(contrast, 90.0)
-            if not np.isfinite(reference) or reference <= 1e-9:
-                return 0.0
-            return float(np.mean(contrast >= 0.15 * reference))
-
-        def _search(ratios, thetas, band):
-            best = (np.inf, 1.0, 0.0)
-            for th in thetas:
-                for rat in ratios:
-                    s = _robust_score(
-                        (1.0, float(rat), float(th)), band, fit_angles
-                    )
-                    if s < best[0]:
-                        best = (s, float(rat), float(th))
-            return best
-
-        # 2. Fit every candidate, clip refinement to the declared search range, and
-        #    validate on held-out angular blocks.
-        coarse_ratios = np.linspace(ratio_range[0], ratio_range[1], n_ratio)
-        coarse_thetas = np.linspace(0.0, 180.0, n_theta, endpoint=False)
-        diagnostics = []
-        ratio_step = (
-            (ratio_range[1] - ratio_range[0]) / max(n_ratio - 1, 1)
-        )
-        for band in candidate_bands:
-            best = _search(coarse_ratios, coarse_thetas, band)
-            if refine:
-                _, rat0, th0 = best
-                dth = 180.0 / n_theta
-                fine_ratios = np.unique(np.clip(
-                    np.linspace(rat0 - ratio_step, rat0 + ratio_step, 11),
-                    ratio_range[0],
-                    ratio_range[1],
-                ))
-                fine_thetas = (
-                    np.linspace(th0 - dth, th0 + dth, 11) % 180.0
-                )
-                fine = _search(fine_ratios, fine_thetas, band)
-                best = min(best, fine, key=lambda item: item[0])
-            fit_score, ratio, theta = best
-            circle_validation = _robust_score(
-                (1.0, 1.0, 0.0), band, validation_angles
-            )
-            ellipse_validation = _robust_score(
-                (1.0, ratio, theta), band, validation_angles
-            )
-            improvement = (
-                (circle_validation - ellipse_validation)
-                / max(abs(circle_validation), 1e-9)
-            )
-            raw_score = _raw_score((1.0, ratio, theta), band)
-            coverage = _angular_coverage((1.0, ratio, theta), band)
-            boundary_limited = (
-                ratio <= ratio_range[0] + 0.25 * ratio_step
-                or ratio >= ratio_range[1] - 0.25 * ratio_step
-            )
-            accepted = (
-                np.isfinite(fit_score)
-                and improvement >= min_fit_improvement
-                and raw_score <= max_fit_score
-                and coverage >= min_angular_coverage
-                and not boundary_limited
-            )
-            diagnostics.append({
-                "band": (float(band[0]), float(band[1])),
-                "r0": float(band[2]),
-                "ratio_b_over_a": float(ratio),
-                "theta_deg": float(theta % 180.0),
-                "fit_score": float(fit_score),
-                "raw_score": float(raw_score),
-                "validation_improvement": float(improvement),
-                "angular_coverage": float(coverage),
-                "boundary_limited": bool(boundary_limited),
-                "accepted": bool(accepted),
-            })
-
-        accepted_candidates = [item for item in diagnostics if item["accepted"]]
-        if accepted_candidates:
-            selected = min(
-                accepted_candidates,
-                key=lambda item: (
-                    item["raw_score"],
-                    -item["validation_improvement"],
-                ),
-            )
-            fit_accepted = True
-        else:
-            selected = min(
-                diagnostics,
-                key=lambda item: (
-                    item["boundary_limited"],
-                    item["raw_score"],
-                    -item["validation_improvement"],
-                ),
-            )
-            fit_accepted = False
-
-        radial_min, radial_max = selected["band"]
-        r0 = selected["r0"]
-        ratio = selected["ratio_b_over_a"] if fit_accepted else 1.0
-        theta_deg = selected["theta_deg"] if fit_accepted else 0.0
-        score = selected["raw_score"]
-
-        # 3. Normalise (a, b) to the selected ring radius; only b/a and theta are
-        #    identifiable. A rejected fit deliberately becomes a circular correction.
-        a_axis, b_axis = r0, r0 * ratio
-        # Canonicalise so a is the MAJOR semi-axis (a/b >= 1): the (a, b, theta) and
-        # (b, a, theta+90) parametrisations describe the same ellipse, so pick the one
-        # with a >= b for an unambiguous a/b >= 1 readout.
-        if b_axis > a_axis:
-            a_axis, b_axis = b_axis, a_axis
-            theta_deg += 90.0
-        theta_deg = float(theta_deg % 180.0)
-        self.ellipse_fit_diagnostics = {
-            "method": "angular_variance",
-            "accepted": fit_accepted,
-            "selected": selected,
-            "candidates": diagnostics,
-            "explicit_band": explicit_band,
-            "center_initial": tuple(float(v) for v in center),
-            "center_refined": tuple(float(v) for v in center),
-            "rejection_reasons": [],
-            "quality_thresholds": {
-                "min_fit_improvement": float(min_fit_improvement),
-                "max_fit_score": float(max_fit_score),
-                "min_angular_coverage": float(min_angular_coverage),
-                "ratio_range": tuple(float(v) for v in ratio_range),
-            },
-        }
-        if not fit_accepted:
-            reasons = []
-            if selected["boundary_limited"]:
-                reasons.append("ratio search boundary")
-            if selected["validation_improvement"] < min_fit_improvement:
-                reasons.append(
-                    f"held-out improvement {selected['validation_improvement']:.3g}"
-                )
-            if selected["raw_score"] > max_fit_score:
-                reasons.append(f"raw score {selected['raw_score']:.3g}")
-            if selected["angular_coverage"] < min_angular_coverage:
-                reasons.append(
-                    f"angular coverage {selected['angular_coverage']:.1%}"
-                )
-            message = (
-                "Ellipse fit rejected; using a circular correction"
-                + (f" ({', '.join(reasons)})." if reasons else ".")
-            )
-            self.ellipse_fit_diagnostics["rejection_reasons"] = reasons
-            warnings.warn(message, RuntimeWarning, stacklevel=2)
-        if verbose:
-            bands_text = ", ".join(
-                f"{item['r0']:.1f}" for item in diagnostics
-            )
-            print(
-                f"  ellipse ring candidates: r0=[{bands_text}] px; "
-                f"selected band=[{radial_min:.1f}, {radial_max:.1f}] px"
-            )
-            print(
-                f"  ellipse ring fit: {'accepted' if fit_accepted else 'rejected'} "
-                f"a/b={a_axis / b_axis:.4f} theta={theta_deg:.2f} deg "
-                f"(score={score:.4g}, held-out improvement="
-                f"{selected['validation_improvement']:.2%}, "
-                f"coverage={selected['angular_coverage']:.1%})"
-            )
-
-        if show:
-            import matplotlib.pyplot as plt
-
-            circ = _polar(dp, (1.0, 1.0, 0.0), radial_min, radial_max)
-            corr = _polar(
-                dp, (a_axis, b_axis, theta_deg), radial_min, radial_max
-            )
-            fig, axes = plt.subplots(1, 3, figsize=(13, 4))
-            axes[0].imshow(dp, cmap="magma")
-            axes[0].plot([origin[1]], [origin[0]], "c+", ms=10)
-            axes[0].set_title("centered mean DP")
-            axes[1].imshow(circ, aspect="auto", cmap="magma")
-            axes[1].set_title("polar: circular (before)")
-            axes[2].imshow(corr, aspect="auto", cmap="magma")
-            axes[2].set_title(f"polar: ellipse-corrected\na/b={a_axis / b_axis:.4f}, θ={theta_deg:.1f}°")
-            for ax in axes[1:]:
-                ax.set_xlabel("radius (band)")
-                ax.set_ylabel("φ bin")
-            plt.tight_layout()
-            plt.show()
-
-        return float(a_axis), float(b_axis), float(theta_deg), (float(radial_min), float(radial_max))
-
-    def _fit_ellipse_from_ridge(
-        self,
-        dp,
-        center,
-        *,
-        radial_min=None,
-        radial_max=None,
-        radial_step=1.0,
-        num_annular_bins=180,
-        ratio_range=(0.85, 1.0),
-        center_search_radius=2.5,
-        max_ring_candidates=3,
-        min_angular_coverage=0.55,
-        min_validation_improvement=0.05,
-        max_validation_residual=2.5,
-        device="cpu",
-        show=False,
-        verbose=False,
-    ):
-        """Jointly refine ring center and ellipticity from a robust radial ridge.
-
-        The diffuse-ring radius is measured independently at each azimuth after
-        log compression and hot-spot clipping. A first/second-harmonic model
-        initializes center and ellipse terms, followed by bounded robust geometric
-        least squares. Fits are accepted only when held-out azimuthal sectors improve
-        over an independently refined circle.
-        """
-        from scipy.optimize import least_squares
-        from quantem.diffraction.polar_transform import polar_transform
-
-        dp = np.asarray(dp, dtype=float)
-        origin = np.asarray(center, dtype=float)
-        qy, qx = dp.shape
-        r_hi = float(min(qy, qx) / 2.0 - 1.0)
-        fit_dp = np.log1p(np.clip(dp, 0.0, None))
-        finite = fit_dp[np.isfinite(fit_dp)]
-        if finite.size:
-            fit_dp = np.minimum(fit_dp, np.percentile(finite, 99.5))
-
-        def polar_at(image, candidate_center, rmin, rmax):
-            return np.asarray(
-                polar_transform(
-                    image,
-                    origin_array=np.asarray(candidate_center, dtype=float),
-                    ellipse_params=(1.0, 1.0, 0.0),
-                    num_annular_bins=num_annular_bins,
-                    radial_min=float(rmin),
-                    radial_max=float(rmax),
-                    radial_step=radial_step,
-                    scan_pos=(0, 0),
-                    device=device,
-                    show_progress=False,
-                ),
-                dtype=float,
-            )
-
-        # Candidate rings from an angular median profile; sparse Bragg spots largely
-        # disappear in the median rather than becoming the selected calibration ring.
-        explicit_band = radial_min is not None and radial_max is not None
-        if explicit_band:
-            bands = [(
-                float(radial_min),
-                float(radial_max),
-                0.5 * (float(radial_min) + float(radial_max)),
-            )]
-        else:
-            full = polar_at(fit_dp, origin, 0.0, r_hi)
-            profile = gaussian_filter1d(np.median(full, axis=0), 2.0)
-            r_axis = np.arange(profile.size, dtype=float) * radial_step
-            exclude = max(6.0, 0.06 * r_hi)
-            start = int(np.ceil(exclude / radial_step))
-            prominence = max(1e-9, 0.03 * np.ptp(profile[start:]))
-            indices, properties = find_peaks(
-                profile,
-                prominence=prominence,
-                distance=max(3, int(round(6.0 / radial_step))),
-            )
-            valid_peaks = (
-                (indices >= start) & (r_axis[indices] <= 0.92 * r_hi)
-            )
-            indices = indices[valid_peaks]
-            prominences = properties["prominences"][valid_peaks]
-            if not indices.size:
-                indices = np.asarray([start + np.argmax(profile[start:])])
-                prominences = np.ones(1)
-            order = np.argsort(prominences)[::-1][:max_ring_candidates]
-            bands = []
-            for index in indices[order]:
-                r0 = float(r_axis[index])
-                half = max(6.0, 0.20 * r0)
-                low = max(exclude, r0 - half) if radial_min is None else float(radial_min)
-                high = min(r_hi, r0 + half) if radial_max is None else float(radial_max)
-                if high > low and not any(abs(r0 - old[2]) < 3.0 for old in bands):
-                    bands.append((low, high, r0))
-            if not bands:
-                fallback_r0 = float(np.clip(r_axis[start], exclude, 0.92 * r_hi))
-                fallback_half = max(6.0, 0.20 * fallback_r0)
-                bands = [(
-                    max(exclude, fallback_r0 - fallback_half),
-                    min(r_hi, fallback_r0 + fallback_half),
-                    fallback_r0,
-                )]
-
-        phi = np.linspace(0.0, 2.0 * np.pi, num_annular_bins, endpoint=False)
-        block_fit = (np.arange(num_annular_bins) // 6) % 2 == 0
-
-        def extract_ridge(band):
-            polar = polar_at(fit_dp, origin, band[0], band[1])
-            polar = np.minimum(
-                polar, np.percentile(polar, 90.0, axis=0, keepdims=True)
-            )
-            smooth = gaussian_filter1d(polar, 1.25, axis=1, mode="nearest")
-            baseline = np.percentile(smooth, 20.0, axis=1, keepdims=True)
-            signal = np.clip(smooth - baseline, 0.0, None)
-            peak_index = np.argmax(signal, axis=1)
-            ridge = np.empty(num_annular_bins, dtype=float)
-            confidence = np.empty(num_annular_bins, dtype=float)
-            for index in range(num_annular_bins):
-                lo = max(0, peak_index[index] - 2)
-                hi = min(signal.shape[1], peak_index[index] + 3)
-                weights = signal[index, lo:hi]
-                bins = np.arange(lo, hi, dtype=float)
-                ridge[index] = (
-                    np.average(bins, weights=weights)
-                    if weights.sum() > 1e-12
-                    else float(peak_index[index])
-                )
-                noise = (
-                    1.4826 * np.median(np.abs(np.diff(smooth[index])))
-                    + 1e-9
-                )
-                confidence[index] = signal[index, peak_index[index]] / noise
-            ridge = band[0] + ridge * radial_step
-            valid = np.isfinite(ridge) & (confidence >= 2.0)
-            if np.count_nonzero(valid) >= 12:
-                design = np.column_stack([
-                    np.ones(num_annular_bins),
-                    np.cos(phi),
-                    np.sin(phi),
-                    np.cos(2 * phi),
-                    np.sin(2 * phi),
-                ])
-                weights = np.clip(confidence / 10.0, 0.05, 1.0)
-                beta = np.zeros(5)
-                for _ in range(5):
-                    use = valid & np.isfinite(weights)
-                    root_weight = np.sqrt(weights[use])
-                    beta = np.linalg.lstsq(
-                        design[use] * root_weight[:, None],
-                        ridge[use] * root_weight,
-                        rcond=None,
-                    )[0]
-                    residual = ridge - design @ beta
-                    scale = 1.4826 * np.median(
-                        np.abs(residual[use] - np.median(residual[use]))
-                    ) + 1e-6
-                    robust = np.minimum(1.0, 1.5 * scale / (np.abs(residual) + 1e-9))
-                    weights = np.clip(confidence / 10.0, 0.05, 1.0) * robust
-                valid &= np.abs(ridge - design @ beta) <= max(3.0, 4.0 * scale)
-            else:
-                beta = np.asarray([np.nan] * 5)
-                weights = np.zeros_like(confidence)
-            return ridge, confidence, valid, beta, weights
-
-        def ellipse_residual(parameters, x, y, weights):
-            cy, cx, axis_a, ratio, theta = parameters
-            dx, dy = x - cx, y - cy
-            cosine, sine = np.cos(theta), np.sin(theta)
-            major = dx * cosine + dy * sine
-            minor = -dx * sine + dy * cosine
-            geometric = (
-                np.sqrt(
-                    (major / axis_a) ** 2
-                    + (minor / (axis_a * ratio)) ** 2
-                )
-                - 1.0
-            ) * axis_a
-            return geometric * np.sqrt(np.clip(weights, 1e-3, None))
-
-        def circle_residual(parameters, x, y, weights):
-            cy, cx, radius = parameters
-            return (
-                np.hypot(x - cx, y - cy) - radius
-            ) * np.sqrt(np.clip(weights, 1e-3, None))
-
-        evaluated = []
-        for band in bands:
-            ridge, confidence, valid, harmonic, weights = extract_ridge(band)
-            coverage = float(np.mean(valid))
-            if np.count_nonzero(valid) < 12:
-                evaluated.append({
-                    "band": tuple(float(v) for v in band[:2]),
-                    "r0": float(band[2]),
-                    "accepted": False,
-                    "angular_coverage": coverage,
-                    "rejection_reasons": ["insufficient ridge points"],
-                })
-                continue
-
-            x = origin[1] + ridge * np.cos(phi)
-            y = origin[0] + ridge * np.sin(phi)
-            fit = valid & block_fit
-            validate = valid & ~block_fit
-            if np.count_nonzero(validate) < 6:
-                fit = valid
-                validate = valid
-
-            r0, c1, s1, c2, s2 = harmonic
-            center_initial = np.asarray([
-                origin[0] + np.clip(s1, -center_search_radius, center_search_radius),
-                origin[1] + np.clip(c1, -center_search_radius, center_search_radius),
-            ])
-            second = float(np.hypot(c2, s2))
-            ratio_initial = np.clip(
-                (max(r0, 1.0) - second) / (max(r0, 1.0) + second),
-                ratio_range[0],
-                ratio_range[1],
-            )
-            theta_initial = 0.5 * np.arctan2(s2, c2)
-            center_low = origin - center_search_radius
-            center_high = origin + center_search_radius
-            axis_low = max(2.0, 0.65 * band[0])
-            axis_high = min(r_hi * 1.5, 1.45 * band[1])
-
-            circle = least_squares(
-                circle_residual,
-                [*center_initial, np.median(ridge[fit])],
-                args=(x[fit], y[fit], weights[fit]),
-                bounds=([
-                    center_low[0], center_low[1], axis_low
-                ], [
-                    center_high[0], center_high[1], axis_high
-                ]),
-                loss="soft_l1",
-                f_scale=1.0,
-            )
-            ellipse = least_squares(
-                ellipse_residual,
-                [
-                    *center_initial,
-                    np.clip(np.max(ridge[fit]), axis_low, axis_high),
-                    ratio_initial,
-                    theta_initial,
-                ],
-                args=(x[fit], y[fit], weights[fit]),
-                bounds=([
-                    center_low[0], center_low[1], axis_low,
-                    ratio_range[0], -np.pi,
-                ], [
-                    center_high[0], center_high[1], axis_high,
-                    ratio_range[1], np.pi,
-                ]),
-                loss="soft_l1",
-                f_scale=1.0,
-            )
-            circle_validation = np.median(np.abs(circle_residual(
-                circle.x, x[validate], y[validate], np.ones(np.count_nonzero(validate))
-            )))
-            ellipse_validation = np.median(np.abs(ellipse_residual(
-                ellipse.x, x[validate], y[validate], np.ones(np.count_nonzero(validate))
-            )))
-            improvement = float(
-                (circle_validation - ellipse_validation)
-                / max(circle_validation, 1e-9)
-            )
-            center_boundary = bool(np.any(
-                np.isclose(ellipse.x[:2], center_low, atol=0.05)
-                | np.isclose(ellipse.x[:2], center_high, atol=0.05)
-            ))
-            ratio_boundary = bool(
-                ellipse.x[3] <= ratio_range[0] + 0.005
-                or ellipse.x[3] >= ratio_range[1] - 0.001
-            )
-            reasons = []
-            if coverage < min_angular_coverage:
-                reasons.append(f"angular coverage {coverage:.1%}")
-            if improvement < min_validation_improvement:
-                reasons.append(f"held-out improvement {improvement:.2%}")
-            if ellipse_validation > max_validation_residual:
-                reasons.append(
-                    f"held-out residual {ellipse_validation:.2f} px"
-                )
-            if center_boundary:
-                reasons.append("center search boundary")
-            if ratio_boundary:
-                reasons.append("ratio search boundary")
-            if not reasons:
-                ellipse = least_squares(
-                    ellipse_residual,
-                    ellipse.x,
-                    args=(x[valid], y[valid], weights[valid]),
-                    bounds=([
-                        center_low[0], center_low[1], axis_low,
-                        ratio_range[0], -np.pi,
-                    ], [
-                        center_high[0], center_high[1], axis_high,
-                        ratio_range[1], np.pi,
-                    ]),
-                    loss="soft_l1",
-                    f_scale=1.0,
-                )
-                center_boundary = bool(np.any(
-                    np.isclose(ellipse.x[:2], center_low, atol=0.05)
-                    | np.isclose(ellipse.x[:2], center_high, atol=0.05)
-                ))
-                ratio_boundary = bool(
-                    ellipse.x[3] <= ratio_range[0] + 0.005
-                    or ellipse.x[3] >= ratio_range[1] - 0.001
-                )
-                if center_boundary:
-                    reasons.append("center search boundary after full refit")
-                if ratio_boundary:
-                    reasons.append("ratio search boundary after full refit")
-            evaluated.append({
-                "band": tuple(float(v) for v in band[:2]),
-                "r0": float(band[2]),
-                "accepted": not reasons,
-                "angular_coverage": coverage,
-                "center_initial": tuple(float(v) for v in center_initial),
-                "center_refined": tuple(float(v) for v in ellipse.x[:2]),
-                "a_pixels": float(ellipse.x[2]),
-                "b_pixels": float(ellipse.x[2] * ellipse.x[3]),
-                "ratio_b_over_a": float(ellipse.x[3]),
-                "theta_deg": float(np.rad2deg(ellipse.x[4]) % 180.0),
-                "circle_validation_residual": float(circle_validation),
-                "ellipse_validation_residual": float(ellipse_validation),
-                "validation_improvement": improvement,
-                "center_boundary": center_boundary,
-                "ratio_boundary": ratio_boundary,
-                "ridge_point_count": int(np.count_nonzero(valid)),
-                "rejection_reasons": reasons,
-                "_ridge": ridge,
-                "_valid": valid,
-                "_x": x,
-                "_y": y,
-            })
-
-        accepted = [item for item in evaluated if item["accepted"]]
-        candidates_with_fit = [item for item in evaluated if "a_pixels" in item]
-        if accepted:
-            selected = min(
-                accepted,
-                key=lambda item: (
-                    item["ellipse_validation_residual"],
-                    -item["validation_improvement"],
-                ),
-            )
-            fit_accepted = True
-        elif candidates_with_fit:
-            selected = min(
-                candidates_with_fit,
-                key=lambda item: item["ellipse_validation_residual"],
-            )
-            fit_accepted = False
-        else:
-            selected = evaluated[0]
-            fit_accepted = False
-
-        public_candidates = [
-            {key: value for key, value in item.items() if not key.startswith("_")}
-            for item in evaluated
-        ]
-        if fit_accepted:
-            a_axis = selected["a_pixels"]
-            b_axis = selected["b_pixels"]
-            theta_deg = selected["theta_deg"]
-            refined_center = selected["center_refined"]
-        else:
-            a_axis = b_axis = selected["r0"]
-            theta_deg = 0.0
-            refined_center = tuple(float(v) for v in origin)
-            reasons = selected.get("rejection_reasons", ["no valid ridge fit"])
-            warnings.warn(
-                "Ridge ellipse fit rejected; using a circular correction "
-                f"({', '.join(reasons)}).",
-                RuntimeWarning,
-                stacklevel=2,
-            )
-
-        selected_public = {
-            key: value for key, value in selected.items() if not key.startswith("_")
-        }
-        self.ellipse_fit_diagnostics = {
-            "method": "ridge",
-            "accepted": fit_accepted,
-            "selected": selected_public,
-            "candidates": public_candidates,
-            "explicit_band": explicit_band,
-            "center_initial": tuple(float(v) for v in origin),
-            "center_refined": tuple(float(v) for v in refined_center),
-            "rejection_reasons": (
-                [] if fit_accepted else selected_public.get("rejection_reasons", [])
-            ),
-        }
-
-        if verbose:
-            print(
-                "  ridge ellipse candidates: "
-                + ", ".join(f"{item['r0']:.1f}" for item in public_candidates)
-                + " px"
-            )
-            print(
-                f"  ridge ellipse fit: {'accepted' if fit_accepted else 'rejected'} "
-                f"a/b={a_axis / b_axis:.4f} theta={theta_deg:.2f} deg "
-                f"center=({refined_center[0]:.2f}, {refined_center[1]:.2f})"
-            )
-
-        if show and "_ridge" in selected:
-            fig, axes = plt.subplots(1, 3, figsize=(13, 4))
-            axes[0].imshow(np.log1p(np.clip(dp, 0, None)), cmap="magma")
-            axes[0].scatter(
-                selected["_x"][selected["_valid"]],
-                selected["_y"][selected["_valid"]],
-                s=5,
-                c="cyan",
-                alpha=0.7,
-                label="ridge inliers",
-            )
-            axes[0].add_patch(Ellipse(
-                (refined_center[1], refined_center[0]),
-                2 * a_axis,
-                2 * b_axis,
-                angle=theta_deg,
-                fill=False,
-                color="lime",
-                linewidth=1.5,
-                label="ridge fit",
-            ))
-            axes[0].legend(fontsize=8)
-            axes[0].set_title("diffuse-ring ridge and robust ellipse")
-            before = polar_at(dp, origin, selected["band"][0], selected["band"][1])
-            after = np.asarray(
-                polar_transform(
-                    dp,
-                    origin_array=np.asarray(refined_center),
-                    ellipse_params=(a_axis, b_axis, theta_deg),
-                    num_annular_bins=num_annular_bins,
-                    radial_min=selected["band"][0],
-                    radial_max=selected["band"][1],
-                    radial_step=radial_step,
-                    scan_pos=(0, 0),
-                    device=device,
-                    show_progress=False,
-                )
-            )
-            axes[1].imshow(before, aspect="auto", cmap="magma")
-            axes[1].set_title("circular polar before")
-            axes[2].imshow(after, aspect="auto", cmap="magma")
-            axes[2].set_title("ridge-refined polar after")
-            fig.tight_layout()
-            plt.show()
-
-        band = selected["band"]
-        return (
-            float(a_axis),
-            float(b_axis),
-            float(theta_deg),
-            (float(band[0]), float(band[1])),
-        )
 
     def resize_data(self, device:str = "cuda:0"):
         print(device)
@@ -3784,148 +2808,21 @@ class BraggPeaksPolymer(AutoSerialize):
             two_fold_symmetry=two_fold_symmetry,
             ellipse_params=center_ellipse_params,
         )
-    
-    def find_central_beams_4d(
-        self,
-        scan_mask: ArrayLike = None,
-        intensity_threshold=0.3,
-        distance_weight=0.5,
-        sampling_radius=2,
-        debug=False,
-        use_tqdm=True,
-        center_method: str = "descent",
-        radial_min: float = 4.0,
-        radial_max: float | None = None,
-        radial_step: float = 1.0,
-        num_annular_bins: int = 180,
-        n_phi: int = 120,
-        kpow: float = 0.0,
-        ellipse_params: tuple[float, float, float] | None = None,
-        center_device: str | None = None,
-        center_batch_size: int = 16,
-        local_margin: int = 40,
-        fallback_to_peaks: bool = True,
-    ):
+
+    def find_central_beams_4d(self, **kwargs):
+        """Locate the central beam at every scan position.
+
+        Thin wrapper over :func:`quantem.diffraction.polar_transform.find_central_beams_4d`,
+        supplying this object's dataset, peaks and device. All keyword arguments are
+        forwarded; see that function for the full set.
         """
-        Fast central beam finding for entire 4D dataset.
-        
-        Parameters:
-        -----------
-        scan_mask : ArrayLike, optional
-            Boolean mask (Ry, Rx) indicating which positions to process
-        use_tqdm : bool
-            Show progress bar
-        
-        Returns:
-        --------
-        centers : ndarray, shape (2, scan_y, scan_x)
-            Center coordinates (y, x) for each scan position
-        """
-        center_method = center_method.lower()
-        if center_method == "peaks":
-            return self._find_central_beams_from_peaks_4d(
-                scan_mask=scan_mask,
-                intensity_threshold=intensity_threshold,
-                distance_weight=distance_weight,
-                sampling_radius=sampling_radius,
-                debug=debug,
-                use_tqdm=use_tqdm,
-            )
-        if center_method not in ("descent", "grid"):
-            raise ValueError(
-                "center_method must be 'descent', 'grid', or 'peaks', "
-                f"got {center_method!r}."
-            )
 
-        scan_y, scan_x, _det_y, _det_x = self.dataset_cartesian.shape
-        if scan_mask is None:
-            scan_mask_arr = np.ones((scan_y, scan_x), dtype=bool)
-        else:
-            scan_mask_arr = np.asarray(scan_mask, dtype=bool)
-            if scan_mask_arr.shape != (scan_y, scan_x):
-                raise ValueError(
-                    f"scan_mask shape {scan_mask_arr.shape} must match {(scan_y, scan_x)}"
-                )
-
-        device = center_device if center_device is not None else self.device
-        try:
-            origins = find_origin_angular_uniformity(
-                self.dataset_cartesian,
-                method=center_method,
-                ellipse_params=ellipse_params,
-                radial_min=radial_min,
-                radial_max=radial_max,
-                radial_step=radial_step,
-                num_annular_bins=num_annular_bins,
-                n_phi=n_phi,
-                kpow=kpow,
-                device=device,
-                batch_size=center_batch_size,
-                local_margin=local_margin,
-            )
-        except Exception as exc:
-            if not fallback_to_peaks:
-                raise
-            warnings.warn(
-                "Angular-uniformity center finding failed; falling back to "
-                f"peak-based central-beam heuristic. Original error: {exc}",
-                stacklevel=2,
-            )
-            return self._find_central_beams_from_peaks_4d(
-                scan_mask=scan_mask_arr,
-                intensity_threshold=intensity_threshold,
-                distance_weight=distance_weight,
-                sampling_radius=sampling_radius,
-                debug=debug,
-                use_tqdm=use_tqdm,
-            )
-
-        if origins.shape != (scan_y, scan_x, 2):
-            raise ValueError(
-                f"Origin finder returned shape {origins.shape}; expected {(scan_y, scan_x, 2)}."
-            )
-        centers = np.moveaxis(np.asarray(origins, dtype=float), -1, 0)
-        centers[:, ~scan_mask_arr] = 0.0
-        return centers
-
-    def _find_central_beams_from_peaks_4d(
-        self,
-        scan_mask: ArrayLike = None,
-        intensity_threshold=0.3,
-        distance_weight=0.5,
-        sampling_radius=2,
-        debug=False,
-        use_tqdm=True,
-    ):
-        """Previous central-beam heuristic based on detected peak locations."""
-        scan_y, scan_x, det_y, det_x = self.dataset_cartesian.shape
-        centers = np.zeros((2, scan_y, scan_x))
-        
-        # Handle scan_mask
-        if scan_mask is None:
-            scan_mask = np.ones((scan_y, scan_x), dtype=bool)
-        else:
-            scan_mask = np.asarray(scan_mask, dtype=bool)
-        
-        iterator = tqdm(range(scan_y), disable=not use_tqdm, desc="Finding centers")
-        
-        for i in iterator:
-            for j in range(scan_x):
-                if not scan_mask[i, j]:
-                    continue
-                if self.peak_coordinates_cartesian[i, j] is None:
-                    print(f"None at i={i}, j={j}")
-                centers[:, i, j] = find_central_beam_from_peaks(
-                    peak_coords=self.peak_coordinates_cartesian[i, j],
-                    peak_intensities=None,
-                    image_shape=(det_y, det_x),
-                    intensity_threshold=intensity_threshold,
-                    distance_weight=distance_weight,
-                    debug=debug,
-                    image=self.dataset_cartesian[i, j].array.squeeze(),
-                    sampling_radius=sampling_radius
-                )
-        return centers
+        return find_central_beams_4d(
+            self.dataset_cartesian,
+            peaks=getattr(self, "peak_coordinates_cartesian", None),
+            default_device=self.device,
+            **kwargs,
+        )
     
     def polar_transform_peaks(
         self,
