+++ tests/diffraction/test_ellipse_ring_fit.py
@@ -2,6 +2,10 @@ import numpy as np
 import pytest
 
 from quantem.diffraction import BraggPeaksPolymer
+from quantem.diffraction.ellipse_fitting import (
+    fit_ellipse_from_ridge,
+    fit_ellipse_from_ring,
+)
 
 
 def _elliptical_ring(
@@ -31,40 +35,39 @@ def _elliptical_ring(
     )
 
 
+def _center(pattern):
+    return ((pattern.shape[0] - 1) / 2, (pattern.shape[1] - 1) / 2)
+
+
 def _fit(pattern, **kwargs):
-    detector = object.__new__(BraggPeaksPolymer)
-    result = detector._fit_ellipse_from_ring(
+    return fit_ellipse_from_ring(
         pattern,
-        ((pattern.shape[0] - 1) / 2, (pattern.shape[1] - 1) / 2),
+        _center(pattern),
         n_ratio=7,
         n_theta=12,
         max_ring_candidates=3,
         **kwargs,
     )
-    return detector, result
 
 
 def _fit_ridge(pattern, **kwargs):
-    detector = object.__new__(BraggPeaksPolymer)
-    center = ((pattern.shape[0] - 1) / 2, (pattern.shape[1] - 1) / 2)
-    result = detector._fit_ellipse_from_ridge(
+    return fit_ellipse_from_ridge(
         pattern,
-        center,
+        _center(pattern),
         num_annular_bins=120,
         **kwargs,
     )
-    return detector, result
 
 
 def test_ring_fit_recovers_synthetic_ellipse():
-    detector, (a_axis, b_axis, theta, band) = _fit(
+    fit = _fit(
         _elliptical_ring(ratio_b_over_a=0.9, theta_deg=35.0)
     )
 
-    assert detector.ellipse_fit_diagnostics["accepted"] is True
-    assert a_axis / b_axis == pytest.approx(1 / 0.9, abs=0.025)
-    assert theta == pytest.approx(35.0, abs=3.0)
-    assert band[0] < 27 < band[1]
+    assert fit.accepted is True
+    assert fit.ratio == pytest.approx(1 / 0.9, abs=0.025)
+    assert fit.theta_deg == pytest.approx(35.0, abs=3.0)
+    assert fit.ring_band[0] < 27 < fit.ring_band[1]
 
 
 def test_sparse_outer_bragg_spots_do_not_select_the_calibration_band():
@@ -77,37 +80,37 @@ def test_sparse_outer_bragg_spots_do_not_select_the_calibration_band():
         column = int(round(cx + 39 * np.cos(angle)))
         pattern[row - 1 : row + 2, column - 1 : column + 2] += 50.0
 
-    detector, (_, _, _, band) = _fit(pattern)
+    fit = _fit(pattern)
 
-    assert detector.ellipse_fit_diagnostics["accepted"] is True
-    assert detector.ellipse_fit_diagnostics["selected"]["r0"] < 32
-    assert band[0] < 26 < band[1]
+    assert fit.accepted is True
+    assert fit.details["selected"]["r0"] < 32
+    assert fit.ring_band[0] < 26 < fit.ring_band[1]
 
 
 def test_boundary_solution_is_rejected_and_refinement_is_clipped():
     pattern = _elliptical_ring(ratio_b_over_a=0.7)
 
     with pytest.warns(RuntimeWarning, match="ratio search boundary"):
-        detector, (a_axis, b_axis, theta, _) = _fit(pattern)
+        fit = _fit(pattern)
 
-    selected = detector.ellipse_fit_diagnostics["selected"]
-    assert detector.ellipse_fit_diagnostics["accepted"] is False
+    selected = fit.details["selected"]
+    assert fit.accepted is False
     assert selected["boundary_limited"] is True
     assert 0.85 <= selected["ratio_b_over_a"] <= 1.18
-    assert (a_axis / b_axis, theta) == pytest.approx((1.0, 0.0))
+    assert (fit.ratio, fit.theta_deg) == pytest.approx((1.0, 0.0))
 
 
 def test_low_information_pattern_falls_back_to_circle():
     with pytest.warns(RuntimeWarning, match="using a circular correction"):
-        detector, (a_axis, b_axis, theta, _) = _fit(np.ones((96, 96)))
+        fit = _fit(np.ones((96, 96)))
 
-    assert detector.ellipse_fit_diagnostics["accepted"] is False
-    assert (a_axis / b_axis, theta) == pytest.approx((1.0, 0.0))
+    assert fit.accepted is False
+    assert (fit.ratio, fit.theta_deg) == pytest.approx((1.0, 0.0))
 
 
 def test_ridge_fit_jointly_recovers_center_and_ellipse():
     offset = (1.2, -0.8)
-    detector, (a_axis, b_axis, theta, band) = _fit_ridge(
+    fit = _fit_ridge(
         _elliptical_ring(
             ratio_b_over_a=0.9,
             theta_deg=35.0,
@@ -116,23 +119,23 @@ def test_ridge_fit_jointly_recovers_center_and_ellipse():
     )
 
     expected_center = np.asarray((47.5, 47.5)) + offset
-    assert detector.ellipse_fit_diagnostics["accepted"] is True
-    assert detector.ellipse_fit_diagnostics["center_refined"] == pytest.approx(
+    assert fit.accepted is True
+    assert fit.center_refined == pytest.approx(
         expected_center, abs=0.25
     )
-    assert a_axis / b_axis == pytest.approx(1 / 0.9, abs=0.03)
-    assert theta == pytest.approx(35.0, abs=3.0)
-    assert band[0] < 27 < band[1]
+    assert fit.ratio == pytest.approx(1 / 0.9, abs=0.03)
+    assert fit.theta_deg == pytest.approx(35.0, abs=3.0)
+    assert fit.ring_band[0] < 27 < fit.ring_band[1]
 
 
 def test_ridge_fit_rejects_out_of_range_ellipse():
     with pytest.warns(RuntimeWarning, match="ratio search boundary"):
-        detector, (a_axis, b_axis, theta, _) = _fit_ridge(
+        fit = _fit_ridge(
             _elliptical_ring(ratio_b_over_a=0.7)
         )
 
-    assert detector.ellipse_fit_diagnostics["accepted"] is False
-    assert (a_axis / b_axis, theta) == pytest.approx((1.0, 0.0))
+    assert fit.accepted is False
+    assert (fit.ratio, fit.theta_deg) == pytest.approx((1.0, 0.0))
 
 
 def test_ridge_fit_ignores_sparse_outer_bragg_spots():
@@ -145,22 +148,22 @@ def test_ridge_fit_ignores_sparse_outer_bragg_spots():
         column = int(round(cx + 39 * np.cos(angle)))
         pattern[row - 1 : row + 2, column - 1 : column + 2] += 50.0
 
-    detector, (_, _, _, band) = _fit_ridge(pattern)
+    fit = _fit_ridge(pattern)
 
-    assert detector.ellipse_fit_diagnostics["accepted"] is True
-    assert detector.ellipse_fit_diagnostics["selected"]["r0"] < 32
-    assert band[0] < 26 < band[1]
+    assert fit.accepted is True
+    assert fit.details["selected"]["r0"] < 32
+    assert fit.ring_band[0] < 26 < fit.ring_band[1]
 
 
 def test_low_information_ridge_falls_back_to_circle():
     with pytest.warns(RuntimeWarning, match="using a circular correction"):
-        detector, (a_axis, b_axis, theta, _) = _fit_ridge(
+        fit = _fit_ridge(
             np.ones((96, 96), dtype=float)
         )
 
-    assert detector.ellipse_fit_diagnostics["method"] == "ridge"
-    assert detector.ellipse_fit_diagnostics["accepted"] is False
-    assert (a_axis / b_axis, theta) == pytest.approx((1.0, 0.0))
+    assert fit.method == "ridge"
+    assert fit.accepted is False
+    assert (fit.ratio, fit.theta_deg) == pytest.approx((1.0, 0.0))
 
 
 def test_preprocess_rejects_unknown_ellipse_fit_method_before_data_access():
@@ -168,3 +171,36 @@ def test_preprocess_rejects_unknown_ellipse_fit_method_before_data_access():
 
     with pytest.raises(ValueError, match="ellipse_fit_method"):
         detector.preprocess(ellipse_fit_method="not-a-fit-method")
+
+
+def test_fit_exposes_the_params_triple_and_legacy_diagnostics_mapping():
+    """``preprocess`` and downstream notebooks depend on both of these shapes.
+
+    ``params`` is the (a, b, theta_deg) triple the polar transforms consume, and
+    ``diagnostics`` reproduces the flat mapping that ``BraggPeaksPolymer`` has
+    always published as ``ellipse_fit_diagnostics``.
+    """
+    fit = _fit(_elliptical_ring(ratio_b_over_a=0.9, theta_deg=35.0))
+
+    assert fit.params == (fit.a, fit.b, fit.theta_deg)
+    assert fit.ratio == pytest.approx(fit.a / fit.b)
+
+    diagnostics = fit.diagnostics
+    assert set(diagnostics) == {
+        "method", "accepted", "selected", "candidates", "explicit_band",
+        "center_initial", "center_refined", "rejection_reasons",
+        "quality_thresholds",
+    }
+    assert diagnostics["method"] == "angular_variance"
+    assert diagnostics["accepted"] is fit.accepted
+    assert diagnostics["center_refined"] == fit.center_refined
+
+
+def test_ridge_diagnostics_mapping_matches_the_ridge_method():
+    fit = _fit_ridge(_elliptical_ring(ratio_b_over_a=0.9, theta_deg=35.0))
+
+    assert set(fit.diagnostics) == {
+        "method", "accepted", "selected", "candidates", "explicit_band",
+        "center_initial", "center_refined", "rejection_reasons",
+    }
+    assert fit.diagnostics["method"] == "ridge"
